fix(rename): Renames depends module to dependable
