+++ README.md
@@ -1,18 +1,18 @@
 <h1 align="center">
-    <strong>depends</strong>
+    <strong>dependable</strong>
 </h1>
 <p align="center">
-    <a href="https://github.com/dmtrs/depends" target="_blank">
-        <img src="https://img.shields.io/github/last-commit/dmtrs/depends" alt="Latest Commit">
+    <a href="https://github.com/dmtrs/dependable" target="_blank">
+        <img src="https://img.shields.io/github/last-commit/dmtrs/dependable" alt="Latest Commit">
     </a>
-        <img src="https://img.shields.io/github/workflow/status/dmtrs/depends/Test">
-        <img src="https://img.shields.io/codecov/c/github/dmtrs/depends">
+        <img src="https://img.shields.io/github/workflow/status/dmtrs/dependable/Test">
+        <img src="https://img.shields.io/codecov/c/github/dmtrs/dependable">
     <br />
-    <a href="https://pypi.org/project/depends" target="_blank">
-        <img src="https://img.shields.io/pypi/v/depends" alt="Package version">
+    <a href="https://pypi.org/project/dependable" target="_blank">
+        <img src="https://img.shields.io/pypi/v/dependable" alt="Package version">
     </a>
-    <img src="https://img.shields.io/pypi/pyversions/depends">
-    <img src="https://img.shields.io/github/license/dmtrs/depends">
+    <img src="https://img.shields.io/pypi/pyversions/dependable">
+    <img src="https://img.shields.io/github/license/dmtrs/dependable">
 </p>
 
 Dependency injection system extracted from `fastapi`
@@ -21,19 +21,21 @@ Dependency injection system extracted from `fastapi`
 import asyncio
 from random import random
 
-from depends import Depends, dependable
+from dependable import dependant, Depends
 
-@dependable
+@dependant
 async def main(*, choice: int = Depends(random)) -> None:
     print(choice)
 
 asyncio.run(main())
 ```
 
+More on [examples](examples/tick.py)
+
 ## Installation
 
 ``` bash
-poetry add depends # pip install depends
+poetry add dependable # pip install dependable
 ```
 
 ## Python 3.6
@@ -46,11 +48,11 @@ poetry add async-exit-stack async_generator # pip install async-exit-stack async
 ## Development
 
 ```bash
-docker build -t depends .
+docker build -t dependable .
 ```
 
 ```bash
-docker run --rm -v $(pwd):/usr/src/app depends scripts/dev
+docker run --rm -v $(pwd):/usr/src/app dependable scripts/dev
 ```
 
 ## References
