+++ mypy.ini
@@ -16,11 +16,11 @@ implicit_reexport = False
 strict_equality = True
 # --strict end
 
-[mypy-depends.concurrency]
+[mypy-dependable.concurrency]
 warn_unused_ignores = False
 ignore_missing_imports = True
 
-[mypy-depends.tests.*]
+[mypy-dependable.tests.*]
 ignore_missing_imports = True
 check_untyped_defs = True
