+++ setup.cfg
@@ -1,6 +1,6 @@
 [isort]
 profile = black
-known_first_party = depends
+known_first_party = dependable
 
 
 [flake8]
